rename automodsumm.py -> core.py
