+++ docs/plasmapy_sphinx/core.py
